+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.12.4 (2025-11-18)
+## 0.12.4 (2025-11-27)
 
 - Added version to the metadata endpoint ([#169](https://github.com/0xMiden/miden-faucet/pull/169)).